# docs/openapi/app.openapi.yaml
openapi: 3.0.3
info:
  title: 鐪熷ソ鐜╁皬绋嬪簭绔?API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/app/profile/service-center/reviews/{reviewId}/like:
    post:
      summary: 点赞收到的评价
      description: 评价目标用户可对收到的评价执行点赞动作；当前记录行为日志并返回 liked 状态。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 返回 reviewId、liked。
        "404":
          description: 评价不存在或不属于当前用户。
  /api/app/profile/service-center/reviews/{reviewId}/actions:
    post:
      summary: 获取评价管理操作菜单
      description: 返回当前评价可执行动作，例如查看关联局、回复评价、申请平台介入。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 返回 actions，每个动作包含 key、text、route。
        "404":
          description: 评价不存在或不属于当前用户。
...
  /api/app/profile/service-center/reviews:
    get:
      summary: Frontend integration endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unified response payload
        "403":
          description: Unified response payload
        "404":
          description: Unified response payload
        "422":
          description: Unified response payload
...
  /api/app/profile/service-center/reviews/{reviewId}/reply:
    post:
      summary: Frontend integration endpoint
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unified response payload
        "403":
          description: Unified response payload
        "404":
          description: Unified response payload
        "422":
          description: Unified response payload
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer